services/api/requirements.txt
# FastAPI Backend Dependencies
# Using versions with pre-built wheels for Python 3.11+

# Core
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.0.0,<3.0.0

# Database
sqlalchemy[asyncio]>=2.0.25,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.0,<5.0.0

# Utilities
cachetools>=5.3.0,<6.0.0
python-multipart>=0.0.6,<1.0.0

# Redis & Background Jobs
redis>=5.0.0,<6.0.0
celery[redis]>=5.3.0,<6.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.26.0,<1.0.0

# Development
python-dotenv>=1.0.0,<2.0.0

